requirements.txt
datasets
openai
python-dotenv
huggingface_hub